+++ crates/ryft-mlir/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 - Added support for the `gpu` dialect.
+- Added support for the Triton `tt` dialect.
 
 ### Changed